rename rich text components
