+++ modules/@apostrophecms/rich-text-widget/ui/apos/components/AposTiptapButton.vue
@@ -15,7 +15,7 @@
 <script>
 
 export default {
-  name: 'ApostropheTiptapButton',
+  name: 'AposTiptapButton',
   props: {
     name: {
       type: String,
