+++ modules/@apostrophecms/rich-text-widget/ui/apos/components/AposTiptapDivider.vue
@@ -7,7 +7,7 @@
 <script>
 
 export default {
-  name: 'ApostropheTiptapDivider',
+  name: 'AposTiptapDivider',
   props: {
     name: {
       type: String,
