+++ modules/@apostrophecms/rich-text-widget/ui/apos/components/AposTiptapStyles.vue
@@ -24,7 +24,7 @@
 <script>
 
 export default {
-  name: 'ApostropheTiptapStyles',
+  name: 'AposTiptapStyles',
   props: {
     name: {
       type: String,
